+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Enhanced `Get-VCFManager` cmdlet to return the SDDC Manager version in `x.y.z` format.
 - Enhanced `Get-VCFManager` cmdlet to return the SDDC Manager build in `xxxxxxx` format.
 - Added `Set-VCFCredentialAutoRotate` cmdlet to configure or disable credential auto-rotation for a credential managed by SDDC Manager.
+- Fixed `validateJsonInput` function to prevent it from truncating directly passed JSON content.
 
 ## v2.3.0